Fix drag and drop when managing the desktop and showing file icons
In on_rootwin_event() no native event is translated into a GDK event
and stored in the event structure that was passed in. So return the
value indicating that the event was not handled and processing must
continue.

The wrong return code caused any dropping onto a drop destination window
lying over PCManFM's window to be passed to the PCManFM window below,
i.e. the intended drop destination window was being ignored.

(This change might actually fix other strange bugs, too.)